refactor: organize Validate code for reusability [experience-6092]
* slight refactor of Waters API interaction for typechecking
* splits form, success / error display, submit button into new components
