+++ frontend-react/src/components/FileHandlers/FileHandlerButton.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import { Button } from "@trussworks/react-uswds";
+
+type FileHandlerSubmitButtonProps = {
+    submitted: boolean;
+    disabled: boolean;
+    reset: () => void;
+    resetText: string;
+    submitText: string;
+};
+
+export const FileHandlerSubmitButton = ({
+    submitted,
+    disabled,
+    reset,
+    resetText,
+    submitText,
+}: FileHandlerSubmitButtonProps) => {
+    if (submitted) {
+        return (
+            <Button type="button" onClick={reset}>
+                {resetText}
+            </Button>
+        );
+    }
+    return (
+        <Button type="submit" disabled={disabled}>
+            {submitText}
+        </Button>
+    );
+};
